refactor contest folder
