+++ .eslintrc.js
@@ -16,6 +16,6 @@ module.exports = {
   rules: {
     "linebreak-style": ["error", "unix"],
     semi: ["error", "always"],
-    "no-unused-vars": 0,
+    "no-unused-vars": ["warn"],
   },
 };
