--- api/v1/controllers/contests/index.js
@@ -1,13 +1,0 @@
-const getAllContests = require("./get-all-contests");
-const createContest = require("./create-contest");
-const getContest = require("./get-contest");
-const updateContest = require("./update-contest");
-const deleteContest = require("./delete-contest");
-
-module.exports = {
-  getAllContests,
-  createContest,
-  getContest,
-  updateContest,
-  deleteContest,
-};
