+++ api/v1/models/index.js
@@ -1,0 +1,38 @@
+'use strict';
+require("dotenv").config();
+
+const Sequelize = require("sequelize");
+
+var fs        = require('fs');
+var path      = require('path');
+var basename  = path.basename(module.filename);
+var env       = process.env.NODE_ENV || 'development';
+var config    = require('./../config/config.js')[env];
+var db        = {};
+
+var sequelize;
+if (config.use_env_variable) {
+  sequelize = new Sequelize(process.env[config.use_env_variable]);
+} else {
+  sequelize = new Sequelize(config.database, config.username, config.password, config);
+}
+
+fs
+  .readdirSync(__dirname)
+  .filter(function(file) {
+    return (file.indexOf('.') !== 0) && (file !== basename) &&     (file.slice(-3) === '.js');
+  })
+  .forEach(function(file) {
+    const model = require(path.join(__dirname, file))(sequelize, Sequelize.DataTypes);
+    db[model.name] = model;
+ });
+
+Object.keys(db).forEach(function(modelName) {
+  if (db[modelName].associate) {
+    db[modelName].associate(db);
+  }
+});
+
+module.exports = {
+  sequelize
+};
