+++ api/v1/utils/contests.js
@@ -1,22 +1,11 @@
-const { validationResult } = require("express-validator");
-
-const { ERRORS } = require("../../constants");
-const { LogicError } = require("../../utils/error-classes");
-
-function dateToTimestamp(date) {
-  return date.getTime() / 1000;
-}
-
-function timestampToDate(timestamp) {
-  return new Date(timestamp * 1000);
-}
+const utils = require("./index");
 
 function formatContest(contest) {
   const materials = contest.materials || {};
   return {
     contest_name: contest.contest_name,
     contest_title: contest.contest_title,
-    start_time: dateToTimestamp(contest.start_time),
+    start_time: utils.dateToTimestamp(contest.start_time),
     duration: contest.duration,
     total_participation: 0, // TODO: fix this after Participations table is done
     can_enter: contest.can_enter,
@@ -31,19 +20,6 @@ function formatContest(contest) {
   };
 }
 
-// throw errors if validation error found
-function validationMiddleware(req, res, next) {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    throw new LogicError({ ...ERRORS.VALIDATION_FAILED, data: errors });
-  } else {
-    return next();
-  }
-}
-
 module.exports = {
-  dateToTimestamp,
-  timestampToDate,
   formatContest,
-  validationMiddleware,
 };
